move managerbehavior to separate file
