+++ src/rsrc_mgr/manager.h
@@ -1,5 +1,5 @@
-#ifndef _RESOURCE_MANAGER_WORK_H
-#define _RESOURCE_MANAGER_WORK_H
+#ifndef _RESOURCE_MANAGER_MANAGER_H
+#define _RESOURCE_MANAGER_MANAGER_H
 /*
 Copyright (c) 2015, Los Alamos National Security, LLC
 All rights reserved.
@@ -59,40 +59,8 @@ https://github.com/jti-lanl/aws4c.
 GNU licenses can be found at http://www.gnu.org/licenses/.
 */
 
-#include "rsrc_mgr/consts.h"
-#include "rsrc_mgr/resourceprocessing.h"
 #include "rsrc_mgr/rmanstate.h"
 
-typedef enum {
-   RLOG_WORK,      // request to process an existing resource log (either previous dry-run or dead run pickup)
-   NS_WORK,        // request to process a portion of a NS
-   COMPLETE_WORK,  // request to complete outstanding work (quiesce all threads and close all streams)
-   TERMINATE_WORK, // request to terminate the rank
-   ABORT_WORK      // request to abort all processing and terminate
-} worktype;
-
-typedef struct {
-   worktype  type;
-   // NS target info
-   size_t    nsindex;
-   size_t    refdist;
-   // Log target info
-   char      iteration[ITERATION_STRING_LEN];
-   size_t    ranknum;
-} workrequest;
-
-typedef struct {
-   workrequest request;
-   // Work results
-   char                 haveinfo;
-   streamwalker_report  report;
-   operation_summary    summary;
-   char                 errorlog;
-   char                 fatalerror;
-   char                 errorstr[MAX_ERROR_BUFFER];
-} workresponse;
-
-int handlerequest(rmanstate* rman, workrequest* request, workresponse* response);
-int handleresponse(rmanstate* rman, size_t ranknum, workresponse* response, workrequest* request);
+int managerbehavior(rmanstate* rman);
 
 #endif
